Refactor: transactional style, 의존성 주입통일
