+++ src/main/java/com/capstone/smutaxi/entity/Message.java
@@ -30,6 +30,4 @@ public class Message {
     @JoinColumn(name = "chat_room_id")
     private ChatRoom chatRoom;
 
-    public Message() {
-    }
 }
