+++ src/main/java/com/capstone/smutaxi/service/auth/EmailService.java
@@ -5,6 +5,7 @@ import com.capstone.smutaxi.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.mail.MessagingException;
@@ -13,6 +14,7 @@ import java.util.concurrent.ThreadLocalRandom;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class EmailService {
 
     private final JavaMailSender javaMailSender;
