+++ src/main/java/com/capstone/smutaxi/utils/ChatRoomInitializer.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * 이 클래스는 스프링 부트 애플리케이션이 시작될때 run 메서드가 실행되는데
